app/requirements.txt
werkzeug>=3.0.1
python-dotenv==1.0.1
Flask-SQLAlchemy
SQLAlchemy
Flask==3.0.3
flask-marshmallow==0.15.0
marshmallow==3.20.1
azure-monitor-opentelemetry==1.0.0
Flask-Migrate==4.0.4
psycopg2-binary==2.9.7
passlib==1.7.4
sqlalchemy-easy-softdelete==0.8.3
Flask-jwt-extended==4.6.0
Flask-Caching>=2.0.2
redis>=5.0.0
gunicorn==21.0.0
Flask-Limiter[redis]==3.5.0
sentry-sdk[flask]==1.40.6
weasyprint
python-pushover
Flask-APScheduler
pytz
wheel
google-cloud-dialogflow==2.24.0
google-auth==2.20.0
boto3==1.34.84
requests
